% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plots_dist.R
\name{get_plots_dist}
\alias{get_plots_dist}
\title{Function 2}
\usage{
get_plots_dist(
  dname,
  n,
  mean = 0,
  sd = 1,
  df = n - 2,
  dfn = 2,
  dfd = n - 2,
  size = 10,
  p = 0.1,
  seed = FALSE,
  bins = 10
)
}
\arguments{
\item{dname}{name of the distribution (normal, t, F, Binomial)}

\item{n}{number of samples}

\item{mean}{vector of means for normal distribution}

\item{sd}{vector of standard deviations for normal distribution}

\item{df}{degrees of freedom for t distribution (> 0, maybe non-integer). df = Inf is allowed}

\item{dfn}{degrees of freedom for the numerator of the f distribution (> 0, maybe non-integer). df = Inf is allowed}

\item{dfd}{degrees of freedom for the denominator of the f distribution (> 0, maybe non-integer). df = Inf is allowed}

\item{size}{number of trials for the Binomial distribution (zero or more)}

\item{p}{probability of success on each trial for Binomial distribution}

\item{seed}{integer vector for the random number generation in R}

\item{bins}{one of:

1. a vector giving the breakpoints between histogram cells,

2. a function to compute the vector of breakpoints,

3. a single number giving the number of cells for the histogram,

4. a character string naming an algorithm to compute the number of cells (see ‘Details’),

5. a function to compute the number of cells

In the last three cases the number is a suggestion only; as the breakpoints will be set to pretty values, the number is limited to 1e6 (with a warning if it was larger). If breaks is a function, the x vector is supplied to it as the only argument (and the number of breaks is only limited by the amount of available memory). freq
logical; if TRUE, the histogram graphic is a representation of frequencies, the counts component of the result; if FALSE, probability densities, component density, are plotted (so that the histogram has a total area of one). Defaults to TRUE if and only if breaks are equidistant (and probability is not specified)}
}
\value{

}
\description{
Generates histograms for random samples from 4 different distributions
}
