% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples_dist.R
\name{get_samples_dist}
\alias{get_samples_dist}
\title{Function 1}
\usage{
get_samples_dist(
  dname,
  n,
  mean = 0,
  sd = 1,
  df = n - 2,
  dfn = 2,
  dfd = n - 2,
  size = 10,
  p = 0.1,
  seed = FALSE
)
}
\arguments{
\item{dname}{name of the distribution (normal, t, F, Binomial)}

\item{n}{number of samples}

\item{mean}{vector of means for normal distribution}

\item{sd}{vector of standard deviations for normal distribution}

\item{df}{degrees of freedom for t distribution (> 0, maybe non-integer). df = Inf is allowed}

\item{dfn}{degrees of freedom for the numerator of the f distribution (> 0, maybe non-integer). df = Inf is allowed}

\item{dfd}{degrees of freedom for the denominator of the f distribution (> 0, maybe non-integer). df = Inf is allowed}

\item{size}{number of trials for the Binomial distribution (zero or more)}

\item{p}{probability of success on each trial for Binomial distribution}

\item{seed}{integer vector for the random number generation in R}
}
\value{

}
\description{
Generates random samples from 4 different distributions
}
\examples{

}
